FIX: Single instance caching
