+++ pythonwhois/caching/whois_server_cache.dat
@@ -1,0 +1,1 @@
+{'forex': u'whois.nic.forex', 'haus': u'whois.unitedtld.com', 'eu': u'whois.eu', 'luxury': 'whois.nic.luxury', 'test': 'whois.nic.test', 'com': u'whois.verisign-grs.com'}
